agv-raspberry/requirements.txt
# Dependências para o sistema AGV Raspberry Pi

# Framework web
flask==2.3.3
flask-cors==4.0.0

# Comunicação serial
pyserial==3.5

# Visão computacional
opencv-python==4.8.1.78
numpy==1.24.3

# Processamento assíncrono
asyncio-mqtt==0.13.0

# Logging e utilitários
python-dotenv==1.0.0

# GPIO (apenas no Raspberry Pi)
# RPi.GPIO==0.7.1  # Descomente no Raspberry Pi real
# gpiozero==1.6.2  # Descomente no Raspberry Pi real
